+++ database/seeders/DatabaseSeeder.php
@@ -15,6 +15,9 @@ final class DatabaseSeeder extends Seeder
     public function run(): void
     {
         $this->call(KraiteSeeder::class);
-        $this->call(BusinessSeeder::class);
+
+        if (app()->environment('local', 'testing')) {
+            $this->call(BusinessSeeder::class);
+        }
     }
 }